requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
pytz==2023.3
tzlocal==5.2
apscheduler==3.10.1
pillow>=9.5.0
requests==2.31.0
reportlab==4.0.4
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.3
PyPDF2==3.0.1
supabase-py>=1.0.3 
httpx>=0.24.0 
aiohttp>=3.8.0